@@ -1,14 +1,12 @@
 ﻿namespace Moq.AutoMock.Tests.Util
 {
-#pragma warning disable CA1801, CA1812  //is an internal class that is apparently never instantiated
-    internal class With3Parameters
+    public class With3Parameters
     {
-        // ReSharper disable UnusedMember.Global
-        // ReSharper disable UnusedParameter.Local
         public With3Parameters() { }
-        public With3Parameters(IService1 service1) { }
+
+#pragma warning disable CA1801  //Parameter  is never used. Remove the parameter or use it in the method body
+        public With3Parameters(IService1 _) { }
         public With3Parameters(IService1 service1, IService2 service2) { }
-        // ReSharper restore UnusedParameter.Local
-        // ReSharper restore UnusedMember.Global
+#pragma warning restore CA1801
     }
 }
