@@ -1,14 +1,12 @@
 ﻿namespace Moq.AutoMock.Tests.Util
 {
-#pragma warning disable CA1801, CA1812  //is an internal class that is apparently never instantiated
-    internal class WithArrayParameter
+    public class WithArrayParameter
     {
-        // ReSharper disable UnusedMember.Global
-        // ReSharper disable UnusedParameter.Local
         public WithArrayParameter() { }
+
+#pragma warning disable CA1801  //Parameter  is never used. Remove the parameter or use it in the method body
         public WithArrayParameter(string[] array) { }
         public WithArrayParameter(string[] array, string @sealed) { }
-        // ReSharper restore UnusedParameter.Local
-        // ReSharper restore UnusedMember.Global
+#pragma warning restore CA1801
     }
 }
