@@ -103,7 +103,7 @@ namespace Moq.AutoMock
             var mock = GetMock<T>();
 
             if (CastChecker.DoesReturnPrimitive(expression))
-                throw new NotSupportedException("Use the Verify overload that allows specifying TReturn if the setup returns a value type");
+                throw new NotSupportedException(Resources.Strings.VerifyWithValueReturn);
 
             mock.Verify(expression);
         }
@@ -121,7 +121,7 @@ namespace Moq.AutoMock
             var mock = GetMock<T>();
 
             if (CastChecker.DoesReturnPrimitive(expression))
-                throw new NotSupportedException("Use the Verify overload that allows specifying TReturn if the setup returns a value type");
+                throw new NotSupportedException(Resources.Strings.VerifyWithValueReturn);
 
             mock.Verify(expression, times);
         }
@@ -139,7 +139,7 @@ namespace Moq.AutoMock
             var mock = GetMock<T>();
 
             if (CastChecker.DoesReturnPrimitive(expression))
-                throw new NotSupportedException("Use the Verify overload that allows specifying TReturn if the setup returns a value type");
+                throw new NotSupportedException(Resources.Strings.VerifyWithValueReturn);
 
             mock.Verify(expression, times);
         }
@@ -157,7 +157,7 @@ namespace Moq.AutoMock
             var mock = GetMock<T>();
 
             if (CastChecker.DoesReturnPrimitive(expression))
-                throw new NotSupportedException("Use the Verify overload that allows specifying TReturn if the setup returns a value type");
+                throw new NotSupportedException(Resources.Strings.VerifyWithValueReturn);
 
             mock.Verify(expression, failMessage);
         }
@@ -176,7 +176,7 @@ namespace Moq.AutoMock
             var mock = GetMock<T>();
 
             if (CastChecker.DoesReturnPrimitive(expression))
-                throw new NotSupportedException("Use the Verify overload that allows specifying TReturn if the setup returns a value type");
+                throw new NotSupportedException(Resources.Strings.VerifyWithValueReturn);
 
             mock.Verify(expression, times, failMessage);
         }
