@@ -8,7 +8,7 @@ namespace Moq.AutoMock
         public static bool DoesReturnPrimitive<TService>(Expression<Func<TService, object>> expression)
             where TService : class
         {
-            if (expression is null) throw new ArgumentNullException(nameof(expression));
+            _ = expression ?? throw new ArgumentNullException(nameof(expression));
 
             //expression is assumed to be a lambda so the cast here would cause an exception if it was not the expected format
             var lambdaExpression = (LambdaExpression)expression;
