@@ -1,3 +1,4 @@
+using Moq.AutoMock.Extensions;
 using System;
 using System.Linq;
 using System.Reflection;
@@ -8,26 +9,33 @@ namespace Moq.AutoMock
     {
         public static ConstructorInfo SelectCtor(this Type type, Type[] existingTypes, BindingFlags bindingFlags)
         {
-            ConstructorInfo? best = null;
-            foreach (var constructor in type.GetConstructors(bindingFlags))
-            {
-                if (IsBetterChoice(constructor))
-                    best = constructor;
-            }
+            ConstructorInfo? best = type
+                .GetConstructors(bindingFlags)
+                .Where(constructor =>
+                {
+                    var parameters = constructor.GetParameters();
 
-            return best ?? throw new ArgumentException($"Did not find a best constructor for `{type}`", nameof(type));
+                    return parameters.Length is 0 || parameters
+                        .Select(parameter => parameter.ParameterType)
+                        .All(type => existingTypes.Contains(type)
+                            || type.IsMockable()
+                            || type.IsArray);
+                })
+                .Aggregate<ConstructorInfo, ConstructorInfo?>(null, (value, constructor) =>
+                {
+                    if (value is null) return constructor;
+                    return value.GetParameters().Length >= constructor.GetParameters().Length ? value : constructor;
+                });
 
-            bool IsBetterChoice(ConstructorInfo candidate)
-            {
-                if (best == null) return true;
+            return best 
+                ?? Empty(type) 
+                ?? throw new ArgumentException(
+                    $"Did not find a best constructor for `{type}`. If your type has a non-public constructor, set the 'enablePrivate' parameter to true for this AutoMocker method.",
+                    nameof(type));
 
-                if (candidate.GetParameters()
-                             .Where(x => !existingTypes.Contains(x.ParameterType))
-                             .Any(x => x.ParameterType.GetTypeInfo().IsSealed && !x.ParameterType.IsArray))
-                    return false;
-
-                return best.GetParameters().Length < candidate.GetParameters().Length;
-            }
+            static ConstructorInfo Empty(Type type) => type
+                .GetConstructors(BindingFlags.Instance | BindingFlags.NonPublic)
+                .FirstOrDefault(x => x.GetParameters().Length is 0);
         }
     }
 }
