@@ -9,12 +9,12 @@ namespace Moq.AutoMock.Extensions
 {
     internal static class TypeExtensions
     {
-        internal static bool TryCompileGetter(this AutoMocker autoMocker, Type funcType, [NotNullWhen(true)] out Delegate? @delegate)
+        public static bool TryCompileGetter(this AutoMocker autoMocker, Type funcType, [NotNullWhen(true)] out Delegate? @delegate)
         {
             @delegate = null;
             var stInfo = funcType.GetTypeInfo();
             if (!typeof(Delegate).IsAssignableFrom(funcType)
-                || !stInfo.IsGenericType || !(funcType.GetGenericTypeDefinition() is Type td)
+                || !stInfo.IsGenericType || funcType.GetGenericTypeDefinition() is not Type td
                 || td.Namespace != nameof(System) || !Regex.IsMatch(td.Name, $"^{nameof(Func<object>)}\\b"))
                 return false;
 
@@ -33,6 +33,13 @@ namespace Moq.AutoMock.Extensions
 
             @delegate = Expression.Lambda(funcType, call, @params).Compile();
             return @delegate != null;
+        }
+
+        public static bool IsDelegateType(this Type type) => type.BaseType == typeof(MulticastDelegate);
+
+        public static bool IsMockable(this Type type)
+        {
+            return (!type.IsSealed && !type.IsArray) || type.IsDelegateType();
         }
     }
 }
